fix: [rec] could generate duplicate xml id definitions.
This was due to the fact that the original xml id (comming from another
module) was checked for existence in local database... and local
database xmlid would all have been attributed to the current model.

As what matters is what is the final xml id of the record once written
in file, we have to look for the localized xmlid name.